src/ml-pipeline-service/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Async and Concurrency
asyncio-mqtt==0.13.0
aiofiles==23.2.1
aioredis==2.0.1

# Task Queue and Orchestration
celery[redis]==5.3.4
kombu==5.3.4
redis==5.0.1

# Event Streaming
pulsar-client==3.3.0

# Machine Learning and Computer Vision
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.206
opencv-python==4.8.1.78
numpy==1.24.4
Pillow==10.1.0
scikit-learn==1.3.2

# Video Processing
ffmpeg-python==0.2.0

# WebSocket Support
websockets==12.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Data Validation and Serialization
marshmallow==3.20.1
jsonschema==4.20.0

# Database and Storage
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Monitoring and Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-celery==0.42b0
opentelemetry-exporter-jaeger==1.21.0
prometheus-client==0.19.0

# Logging and Configuration
structlog==23.2.0
python-json-logger==2.0.7
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI
factory-boy==3.3.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
click==8.1.7
tqdm==4.66.1
python-multipart==0.0.6
email-validator==2.1.0

# GPU Support (optional)
# Uncomment if using CUDA
# torch==2.1.1+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# torchvision==0.16.1+cu118 -f https://download.pytorch.org/whl/torch_stable.html
